error_category_map/services/dlq-handler/requirements.txt
kafka-python==2.0.2
kubernetes==30.1.0